+++ packages/eslint-plugin/src/rules/prefer-string-starts-ends-with.ts
@@ -47,7 +47,7 @@ export default createRule({
      * Check if a given node is a string.
      * @param node The node to check.
      */
-    function isStringType(node: TSESTree.LeftHandSideExpression): boolean {
+    function isStringType(node: TSESTree.Expression): boolean {
       const objectType = services.getTypeAtLocation(node);
       return getTypeName(checker, objectType) === 'string';
     }