+++ packages/eslint-plugin/src/rules/unbound-method.ts
@@ -236,7 +236,7 @@ export default util.createRule<Options, MessageIds>({
               }
 
               checkMethodAndReport(
-                node,
+                property.key,
                 initTypes.getProperty(property.key.name),
               );
             }