+++ packages/typescript-estree/src/index.ts
@@ -21,9 +21,6 @@ export * from './create-program/getScriptKind';
 export { typescriptVersionIsAtLeast } from './version-check';
 export * from './getModifiers';
 
-// re-export for backwards-compat
-export { visitorKeys } from '@typescript-eslint/visitor-keys';
-
 // note - cannot migrate this to an import statement because it will make TSC copy the package.json to the dist folder
 // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
 export const version: string = require('../package.json').version;