+++ packages/typescript-estree/tests/lib/parse.test.ts
@@ -6,7 +6,6 @@ import * as astConverterModule from '../../src/ast-converter';
 import * as sharedParserUtilsModule from '../../src/create-program/shared';
 import type { TSESTreeOptions } from '../../src/parser-options';
 import { createSnapshotTestBlock } from '../../tools/test-utils';
-import { expectToHaveParserServices } from './test-utils/expectToHaveParserServices';
 
 const FIXTURES_DIR = join(__dirname, '../fixtures/simpleProject');
 
@@ -349,6 +348,7 @@ describe('parseAndGenerateServices', () => {
         let result:
           | parser.ParseAndGenerateServicesResult<typeof config>
           | undefined;
+        // eslint-disable-next-line jest/valid-expect
         const exp = expect(() => {
           result = parser.parseAndGenerateServices(code, {
             ...config,
@@ -709,93 +709,6 @@ describe('parseAndGenerateServices', () => {
       const ignore = ['**/ignoreme/**'];
       expect(testParse('ignoreme', ignore)).toThrow();
       expect(testParse('includeme', ignore)).not.toThrow();
-    });
-  });
-
-  describe('moduleResolver', () => {
-    beforeEach(() => {
-      parser.clearCaches();
-    });
-
-    const PROJECT_DIR = resolve(FIXTURES_DIR, '../moduleResolver');
-    const code = `
-      import { something } from '__PLACEHOLDER__';
-
-      something();
-    `;
-    const config: TSESTreeOptions = {
-      comment: true,
-      tokens: true,
-      range: true,
-      loc: true,
-      project: './tsconfig.json',
-      tsconfigRootDir: PROJECT_DIR,
-      filePath: resolve(PROJECT_DIR, 'file.ts'),
-    };
-    const withDeprecatedDefaultProgramConfig: TSESTreeOptions = {
-      ...config,
-      project: './tsconfig.defaultProgram.json',
-      DEPRECATED__createDefaultProgram: true,
-    };
-
-    describe('when file is in the project', () => {
-      it('returns error if __PLACEHOLDER__ can not be resolved', () => {
-        const services = parser.parseAndGenerateServices(code, config).services;
-        expectToHaveParserServices(services);
-        const diagnositcs = services.program.getSemanticDiagnostics();
-        expect(diagnositcs.length).toBeGreaterThan(0);
-        expect(diagnositcs[0].messageText).toBe(
-          "Cannot find module '__PLACEHOLDER__' or its corresponding type declarations.",
-        );
-      });
-
-      it('throws error if moduleResolver can not be found', () => {
-        expect(() =>
-          parser.parseAndGenerateServices(code, {
-            ...config,
-            moduleResolver: resolve(
-              PROJECT_DIR,
-              './this_moduleResolver_does_not_exist.js',
-            ),
-          }),
-        ).toThrowErrorMatchingInlineSnapshot(`
-          "Could not find the provided parserOptions.moduleResolver.
-          Hint: use an absolute path if you are not in control over where the ESLint instance runs."
-        `);
-      });
-
-      it('resolves __PLACEHOLDER__ correctly', () => {
-        const services = parser.parseAndGenerateServices(code, {
-          ...config,
-          moduleResolver: resolve(PROJECT_DIR, './moduleResolver.js'),
-        }).services;
-        expectToHaveParserServices(services);
-        expect(services.program.getSemanticDiagnostics()).toHaveLength(0);
-      });
-    });
-
-    describe('when file is not in the project and DEPRECATED__createDefaultProgram=true', () => {
-      it('returns error because __PLACEHOLDER__ can not be resolved', () => {
-        const services = parser.parseAndGenerateServices(
-          code,
-          withDeprecatedDefaultProgramConfig,
-        ).services;
-        expectToHaveParserServices(services);
-        const diagnositcs = services.program.getSemanticDiagnostics();
-        expect(diagnositcs.length).toBeGreaterThan(0);
-        expect(diagnositcs[0].messageText).toBe(
-          "Cannot find module '__PLACEHOLDER__' or its corresponding type declarations.",
-        );
-      });
-
-      it('resolves __PLACEHOLDER__ correctly', () => {
-        const services = parser.parseAndGenerateServices(code, {
-          ...withDeprecatedDefaultProgramConfig,
-          moduleResolver: resolve(PROJECT_DIR, './moduleResolver.js'),
-        }).services;
-        expectToHaveParserServices(services);
-        expect(services.program.getSemanticDiagnostics()).toHaveLength(0);
-      });
     });
   });
 });